Removes database
KO for database. It has no sense have got database into this kind of app. Design's error produces a lot of work for nothing...

I removed all refer of db and now the app is also more responsive (maybe #13 is not more a bug).
For this it will close #18 that now it is not necessary and underlines the need to implement RxJava #15